plugins: consolidate and collapse some text-based filters
