--- src/_layouts/compress.html
@@ -1,4 +1,0 @@
----
----
-
-{{ content | compress_html }}
