+++ src/_layouts/default.html
@@ -1,5 +1,5 @@
 ---
-layout: compress
+layout: minify_html
 ---
 
 <!DOCTYPE html>
