+++ src/_layouts/minify_html.html
@@ -1,0 +1,4 @@
+---
+---
+
+{{ content | minify_html }}
