+++ src/_plugin_tests/test_text_utils.rb
@@ -2,31 +2,31 @@
 
 require 'test/unit'
 
-require_relative '../_plugins/cleanup_text'
+require_relative '../_plugins/utils/text'
 
-class TextCleanupText < Test::Unit::TestCase
+class TestTextUtils < Test::Unit::TestCase
   def test_it_adds_non_breaking_spaces_after_words
     text = 'Apollo 11 launched in 1969'
     expected = 'Apollo&nbsp;11 launched in 1969'
-    assert_equal(cleanup_whitespace(text), expected)
+    assert_equal(Alexwlchan::TextUtils.add_non_breaking_characters(text), expected)
   end
 
   def test_it_adds_non_breaking_spaces_before_words
     text = 'It takes 2 minutes'
     expected = 'It takes 2&nbsp;minutes'
-    assert_equal(cleanup_whitespace(text), expected)
+    assert_equal(Alexwlchan::TextUtils.add_non_breaking_characters(text), expected)
   end
 
   def test_non_breaking_space_at_start_of_anchor
     text = '<a href="https://kottke.org/">a different set of circles</a>'
     expected = '<a href="https://kottke.org/">a&nbsp;different set of circles</a>'
-    assert_equal(cleanup_whitespace(text), expected)
+    assert_equal(Alexwlchan::TextUtils.add_non_breaking_characters(text), expected)
   end
 
   def test_thinsp_around_times
     text = '100×200'
     expected = '100&#8239;×&#8239;200'
-    assert_equal(cleanup_whitespace(text), expected)
+    assert_equal(Alexwlchan::TextUtils.add_non_breaking_characters(text), expected)
   end
 
   def test_preserves_language_console
@@ -34,25 +34,25 @@ class TextCleanupText < Test::Unit::TestCase
 </code></pre></div></div>'
     expected = '<pre class="lng-console"><code><span class="gp">$</span><span class="w"> </span>youtube-dl <span class="nt">--write-auto-sub</span> <span class="nt">--skip-download</span> <span class="s2">"https://www.youtube.com/watch?v=XyGVRlRyT-E"</span>
 </code></pre>'
-    assert_equal(cleanup_syntax_highlighter_classes(text), expected)
+    assert_equal(Alexwlchan::TextUtils.cleanup_syntax_highlighter_classes(text), expected)
   end
 
   def test_preserves_language_go
     text = '<div class="language-go highlighter-rouge"><div class="highlight"><pre class="highlight"><code><span class="k">package</span> <span class="n">main</span></code></pre></div></div>'
     expected = '<pre class="lng-go"><code><span class="k">package</span> <span class="n">main</span></code></pre>'
-    assert_equal(cleanup_syntax_highlighter_classes(text), expected)
+    assert_equal(Alexwlchan::TextUtils.cleanup_syntax_highlighter_classes(text), expected)
   end
 
   def test_discards_other_language_tags_in_pre
     text = '<div class="language-python highlighter-rouge"><div class="highlight"><pre class="highlight"><code><span class="nx">module</span> <span class="s2">"important_bukkit"</span> <span class="p">{</span></pre></div></div>'
     expected = '<pre><code><span class="nx">module</span> <span class="s2">"important_bukkit"</span> <span class="p">{</span></pre>'
-    assert_equal(cleanup_syntax_highlighter_classes(text), expected)
+    assert_equal(Alexwlchan::TextUtils.cleanup_syntax_highlighter_classes(text), expected)
   end
 
   def test_discards_other_language_tags_in_code
     text = '<code class="language-plaintext highlighter-rouge">number_of_sides</code>'
     expected = '<code>number_of_sides</code>'
-    assert_equal(cleanup_syntax_highlighter_classes(text), expected)
+    assert_equal(Alexwlchan::TextUtils.cleanup_syntax_highlighter_classes(text), expected)
   end
 
   def test_handles_highlight_inside_list
@@ -72,7 +72,7 @@ class TextCleanupText < Test::Unit::TestCase
       </li>
     </ul>
 HTML
-    actual = cleanup_syntax_highlighter_classes(text)
+    actual = Alexwlchan::TextUtils.cleanup_syntax_highlighter_classes(text)
     refute(actual.include?('</div>'))
   end
 end
