+++ src/_plugin_tests/tests.rb
@@ -16,8 +16,8 @@ end
 require_relative 'test_add_utm_source'
 require_relative 'test_atom_feed_filters'
 require_relative 'test_attrs'
-require_relative 'test_cleanup_text'
 require_relative 'test_inline_styles'
 require_relative 'test_picture_utils'
 require_relative 'test_tag_tweet'
+require_relative 'test_text_utils'
 require_relative 'test_tint_colors'
