--- src/_plugins/utils/markdownify_oneline.rb
@@ -1,7 +1,0 @@
-def apply_markdownify_oneline(site, input)
-  site.find_converter_instance(Jekyll::Converters::Markdown)
-      .convert(input)
-      .sub('<p>', '')
-      .sub('</p>', '')
-      .strip
-end
