+++ src/_posts/2019/2019-05-24-first-thoughts-on-trans-inclusion.md
@@ -310,6 +310,6 @@ And if you know of other good ones, please let me know!
 This is just part of a bigger conversation at Wellcome about how we do a better job of trans inclusion.
 A policy is part of that, but there's more we can do -- I'll be thinking about it more over the next few months, and hopefully I'll write about some of it here.
 
-If you're interested in what we're doing, please get in touch -- by [Twitter] or <a href="{{ site.email | encode_mailto }}">email</a>.
+If you're interested in what we're doing, please get in touch -- by [Twitter] or <a href="mailto:alex@alexwlchan.net">email</a>.
 
 [Twitter]: https://twitter.com/alexwlchan
