+++ src/static/style.scss
@@ -13,7 +13,6 @@
 @use "components/dot_list";
 @use "components/footer";
 @use "components/images_wide";
-@use "components/latex";
 @use "components/nav";
 @use "components/pretty_hr";
 @use "components/skip_to_main";
